Add Room database and alarm management features
summery:
restructured the code to align with MVVM,
changed the whole UI to be modern

in Detail:
Introduced Room database integration for alarms, including Alarm, AlarmDao, and AlarmDataBase classes. Added AlarmScheduler for scheduling/cancelling alarms, BootReceiver for rescheduling alarms on boot, and updated AlarmReceiver and AlarmService for database-driven alarm handling. Refactored MainActivity for navigation and ViewModel usage, added adapters and fragments for alarms and rules, and updated AndroidManifest and Gradle dependencies to support new features.